+++ src/app/admin/control/action.ts
@@ -1,13 +1,11 @@
 'use server';
 
-import { getAbsenceList } from '@/database/coffeebean/get';
 import { toggleOrderBlock } from '@/database/coffeebean/patch';
 import {
   crawlAndSaveCategory,
   crawlAndSaveMenu,
 } from '@/database/coffeebean/post';
-import { ServerActionState } from '@/hooks/useServerAction';
-import { Absence, OrderBlock } from '@/type';
+import { OrderBlock } from '@/type';
 
 type CategoryCrawlType =
   | Awaited<ReturnType<typeof crawlAndSaveCategory>>['categoryList']
@@ -48,24 +46,3 @@ export const toggleOrderState = async (): Promise<
     message,
   };
 };
-
-export async function getAbsenceListAction(
-  _: ServerActionState<Absence[]>,
-): Promise<ServerActionState<Absence[]>> {
-  try {
-    const absenceList = await getAbsenceList();
-    if (!absenceList || absenceList.length === 0)
-      return { status: 'success', message: '결석 인원이 없습니다.' };
-    return {
-      status: 'success',
-      message: '성공적으로 불러왔습니다.',
-      data: absenceList,
-    };
-  } catch {
-    return { status: 'error', message: '메뉴를 불러올 수 없습니다.' };
-  }
-}
-
-// export const toggleAbsenceAction = async () => {
-
-// }
